Parse Long to integer

+++ src/main/java/org/example/delegate/SendChangesDelegate.java
@@ -18,8 +18,8 @@ public class SendChangesDelegate implements JavaDelegate {
         JobOpeningInformation jobInfo = new JobOpeningInformation();
         jobInfo.setWBIG_processInstanceID(wbig_processInstanceId);
         jobInfo.setOpening_name((String) execution.getVariable("openingName"));
-        jobInfo.setOpen_spots_initial((Integer) execution.getVariable("openSpots"));
-        jobInfo.setOpen_spots_remaining((Integer) execution.getVariable("openSpotsRemaining"));
+        jobInfo.setOpen_spots_initial(Math.toIntExact((Long)execution.getVariable("openSpots")));
+        jobInfo.setOpen_spots_remaining(Math.toIntExact((Long) execution.getVariable("openSpotsRemaining")));
         jobInfo.setSalary((Double) execution.getVariable("salary"));
         jobInfo.setJob_title((String) execution.getVariable("jobTitle"));
         jobInfo.setJob_description((String) execution.getVariable("jobDescription"));
@@ -28,7 +28,7 @@ public class SendChangesDelegate implements JavaDelegate {
         jobInfo.setDeadline((Date) execution.getVariable("deadline"));
         jobInfo.setRewardPerAcceptance((Double) execution.getVariable("paymentInformationAcceptances"));
         jobInfo.setJob_location((String) execution.getVariable("jobLocation"));
-        jobInfo.setWorking_hours((Integer) execution.getVariable("workingHours"));
+        jobInfo.setWorking_hours(Math.toIntExact((Long) execution.getVariable("workingHours")));
 
         //set endpoint and post object
         String EndpointUrl = new wbigRestEndpoints().getCurrent_URL();
